Refactor all endpoints to attend to session main relationship

+++ src/main/java/com/grglucastr/roshambo/controller/v1/PlayerController.java
@@ -1,49 +1,82 @@
 package com.grglucastr.roshambo.controller.v1;
 
+import com.grglucastr.roshambo.model.Move;
 import com.grglucastr.roshambo.model.Player;
+import com.grglucastr.roshambo.model.RoshamboSession;
+import com.grglucastr.roshambo.model.Strategy;
+import com.grglucastr.roshambo.repository.v1.MoveRepository;
 import com.grglucastr.roshambo.repository.v1.PlayerRepository;
+import com.grglucastr.roshambo.repository.v1.RoshamboSessionRepository;
+import com.grglucastr.roshambo.repository.v1.StrategyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @RestController
-@RequestMapping("/api/v1/players")
-public class PlayerController {
+@RequestMapping("/api/v1/roshambo-sessions/{sessionId}/players")
+public class PlayerController extends BaseController{
 
     private PlayerRepository playerRepository;
+    private MoveRepository moveRepository;
+    private StrategyRepository strategyRepository;
 
     @Autowired
-    public PlayerController(PlayerRepository playerRepository) {
-        this.playerRepository = playerRepository;
+    public PlayerController(RoshamboSessionRepository roshamboRepository) {
+        super(roshamboRepository);
     }
 
     @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<Player>>listAllPlayers(){
-        return ResponseEntity.ok(playerRepository.listAll());
-    }
+    public ResponseEntity<Set<Player>>listAllPlayers(@PathVariable("sessionId") Integer sessionId){
 
-    @GetMapping(value = "/{username}",  produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Player> getPlayer(@PathVariable("username") String username){
-        Optional<Player> player = playerRepository.findByUsername(username);
-
-        return player.map(p -> ResponseEntity.ok(p))
+        Optional<RoshamboSession> optSession = getSession(sessionId);
+        return optSession.map(session -> ResponseEntity.ok(session.getPlayers()))
                 .orElse(ResponseEntity.notFound().build());
     }
 
+    @GetMapping(value = "/{playerId}",  produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Player> getPlayerByPlayerId(@PathVariable("sessionId") Integer sessionId,
+                                                      @PathVariable("playerId") Integer playerId){
+        Optional<RoshamboSession> optSession = getSession(sessionId);
+        if(optSession.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        playerRepository = new PlayerRepository(optSession.get().getPlayers());
+        Optional<Player> optPlayer = playerRepository.findById(playerId);
+        return optPlayer.map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
+    }
+
     @PostMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Player> addNewPlayer(@RequestBody Player newPlayer){
+    public ResponseEntity<Player> addNewPlayer(@PathVariable("sessionId") Integer sessionId,
+                                               @RequestBody Player newPlayer){
+
+        Optional<RoshamboSession> optSession = getSession(sessionId);
+        if(optSession.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        playerRepository = new PlayerRepository(optSession.get().getPlayers());
         Player player = playerRepository.add(newPlayer);
+        optSession.get().getPlayers().add(player);
         return ResponseEntity.status(HttpStatus.CREATED).body(player);
     }
 
-    @PutMapping(value = "/{username}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Player> update(@PathVariable("username")String username, @RequestBody Player player){
-        Optional<Player> foundPlayer = playerRepository.findByUsername(username);
+    @PutMapping(value = "/{playerId}", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Player> update(@PathVariable("sessionId") Integer sessionId,
+                                         @PathVariable("playerId") Integer id,
+                                         @RequestBody Player player){
+
+        Optional<RoshamboSession> optSession = getSession(sessionId);
+        if(optSession.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        playerRepository = new PlayerRepository(optSession.get().getPlayers());
+        Optional<Player> foundPlayer = playerRepository.findById(id);
 
         if(foundPlayer.isEmpty()){
             return ResponseEntity.notFound().build();
@@ -51,16 +84,117 @@ public class PlayerController {
 
         Player updatePlayer = foundPlayer.get();
         updatePlayer.setName(player.getName());
+
         playerRepository.update(updatePlayer);
+        optSession.get().getPlayers().addAll(playerRepository.listAll());
 
         return ResponseEntity.ok(updatePlayer);
     }
 
-    @DeleteMapping( value = "/{id}")
-    public ResponseEntity deletePlayer(@PathVariable("id") Integer id){
+    @DeleteMapping( value = "/{playerId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Player> deletePlayer(@PathVariable("sessionId") Integer sessionId,
+                                               @PathVariable("playerId") Integer id){
+
+        Optional<RoshamboSession> optSession = getSession(sessionId);
+        if(optSession.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        playerRepository = new PlayerRepository(optSession.get().getPlayers());
+        Optional<Player> foundPlayer = playerRepository.findById(id);
+
+        if(foundPlayer.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
         if(playerRepository.removeById(id)){
             return ResponseEntity.noContent().build();
         }
+
+        optSession.get().getPlayers().addAll(playerRepository.listAll());
+
         return ResponseEntity.notFound().build();
     }
+
+    @GetMapping(value = "{playerId}/moves", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Set<Move>> listAllPlayerMOve(@PathVariable("sessionId") Integer sessionId,
+                                                       @PathVariable("playerId") Integer playerId){
+
+        Optional<RoshamboSession> optSession = getSession(sessionId);
+        if(optSession.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        playerRepository = new PlayerRepository(optSession.get().getPlayers());
+        Optional<Player> foundPlayer = playerRepository.findById(playerId);
+
+        if(foundPlayer.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        moveRepository = new MoveRepository(optSession.get().getMoves());
+        Set<Move> playerMoves = moveRepository.findMovesByPlayerId(foundPlayer.get().getId());
+        return ResponseEntity.ok(playerMoves);
+    }
+
+    @PostMapping(value = "{playerId}/moves", consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Move> addPlayerMove(@PathVariable("sessionId") Integer sessionId,
+                                              @PathVariable("playerId") Integer playerId,
+                                              @RequestBody Strategy strategy){
+
+        Optional<RoshamboSession> optSession = getSession(sessionId);
+        if(optSession.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        playerRepository = new PlayerRepository(optSession.get().getPlayers());
+        Optional<Player> foundPlayer = playerRepository.findById(playerId);
+        if(foundPlayer.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        strategyRepository = new StrategyRepository(optSession.get().getStrategies());
+        Optional<Strategy> foundStrategy = strategyRepository.findById(strategy.getId());
+        if(foundStrategy.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        moveRepository = new MoveRepository(optSession.get().getMoves());
+        Move newMove = moveRepository.add(new Move(null, foundPlayer.get(), foundStrategy.get()));
+        optSession.get().getMoves().add(newMove);
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(newMove);
+    }
+
+
+    @DeleteMapping(value = "{playerId}/moves/{moveId}", produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<Move> removePlayerMove(@PathVariable("sessionId") Integer sessionId,
+                                                 @PathVariable("playerId") Integer playerId,
+                                                 @PathVariable("moveId") Integer moveId){
+        Optional<RoshamboSession> optSession = getSession(sessionId);
+        if(optSession.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        playerRepository = new PlayerRepository(optSession.get().getPlayers());
+        Optional<Player> foundPlayer = playerRepository.findById(playerId);
+
+        if(foundPlayer.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        moveRepository = new MoveRepository(optSession.get().getMoves());
+        Optional<Move> foundMove = moveRepository.findById(moveId);
+
+        if(foundPlayer.isEmpty()){
+            return ResponseEntity.notFound().build();
+        }
+
+        moveRepository.removeById(moveId);
+        optSession.get().getMoves().addAll(moveRepository.listAll());
+
+        return ResponseEntity.noContent().build();
+    }
+
+
 }
